# PERF: faster numeric indexes comparisons when self is identical to other 