--- a/pandas/core/indexes/range.py
+++ b/pandas/core/indexes/range.py
@@ -813,11 +813,8 @@ class RangeIndex(Int64Index):
 
     def _cmp_method(self, other, op):
         if isinstance(other, RangeIndex) and self._range == other._range:
-            if op in {operator.eq, operator.le, operator.ge}:
-                return np.ones(len(self), dtype=bool)
-            elif op in {operator.ne, operator.lt, operator.gt}:
-                return np.zeros(len(self), dtype=bool)
-
+            # Both are immutable so if ._range attr. are equal, shortcut is possible
+            return super()._cmp_method(self, op)
         return super()._cmp_method(other, op)
 
     def _arith_method(self, other, op):